Move TLSA record handling into a submodule
