+++ pyproject.toml
@@ -94,9 +94,9 @@ convention = "google"
 
 [tool.ruff.lint.flake8-bugbear]
 extend-immutable-calls = [
-    "desec.__init__.TLSAUsage",
-    "desec.__init__.TLSASelector",
-    "desec.__init__.TLSAMatchType",
+    "desec.tlsa.TLSAUsage",
+    "desec.tlsa.TLSASelector",
+    "desec.tlsa.TLSAMatchType",
 ]
 
 [tool.ruff.lint.flake8-pytest-style]
